hacking and negative tests added
WORKING.
