+++ src/test/resources/authenticated/inventor/negativeBecomeInventor.csv
@@ -1,0 +1,3 @@
+company,statement,info
+, Technology, https://info.com
+
