+++ src/test/resources/authenticated/patron/negativeBecomePatron.csv
@@ -1,0 +1,3 @@
+company,statement,info
+Acme,, https://patronInfo.com
+
